+++ VisualChatGPTStudioShared/Utils/CodeCompletion/CompletionData.cs
@@ -2,6 +2,7 @@
 using ICSharpCode.AvalonEdit.Document;
 using ICSharpCode.AvalonEdit.Editing;
 using System;
+using System.Collections.Generic;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 
@@ -40,24 +41,34 @@ namespace JeffPires.VisualChatGPTStudio.Utils.CodeCompletion
         public double Priority => 0;
 
         /// <summary>
-        /// To store auxiliar data
+        /// Gets the type of the completion item.
         /// </summary>
-        public string Aux { get; private set; }
+        public CompletionItemType CompletionItemType { get; private set; }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="CompletionData"/> class.
+        /// Gets the file path as a string.
         /// </summary>
-        /// <param name="text">The text associated with the completion data.</param>
-        /// <param name="description">A description providing additional information about the completion data.</param>
-        /// <param name="image">An image source representing an icon or visual associated with the completion data.</param>
-        /// <param name="aux">Auxiliar data if necssary.</param>
-        public CompletionData(string text, string description, ImageSource image, string aux)
-        {
-            this.Text = text;
-            this.Description = description;
-            this.Image = image;
-            this.Aux = aux;
-        }
+        public string FilePath { get; private set; }
+
+        /// <summary>
+        /// Gets the name of the class.
+        /// </summary>
+        public string ClassName { get; private set; }
+
+        /// <summary>
+        /// Gets the name of the method.
+        /// </summary>
+        public string MethodName { get; private set; }
+
+        /// <summary>
+        /// Gets the list of method parameter types as strings.
+        /// </summary>
+        public List<string> MethodParameterTypes { get; private set; }
+
+        /// <summary>
+        /// Gets or sets the method signature.
+        /// </summary>
+        public string MethodSignature { get; set; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="CompletionData"/> class with the specified text, description, and image.
@@ -80,6 +91,69 @@ namespace JeffPires.VisualChatGPTStudio.Utils.CodeCompletion
         }
 
         /// <summary>
+        /// Initializes a new instance of the <see cref="CompletionData"/> class.
+        /// </summary>
+        /// <param name="text">The text of the completion item.</param>
+        /// <param name="description">A description of the completion item.</param>
+        /// <param name="image">An image associated with the completion item.</param>
+        /// <param name="completionItemType">The type of the completion item.</param>
+        /// <param name="filePath">The file path where the completion item is located.</param>
+        /// <param name="className">The name of the class associated with the completion item.</param>
+        /// <returns>
+        /// A new instance of the <see cref="CompletionData"/> class.
+        /// </returns>
+        public CompletionData(string text,
+                              string description,
+                              ImageSource image,
+                              CompletionItemType completionItemType,
+                              string filePath,
+                              string className)
+        {
+            this.Text = text;
+            this.Description = description;
+            this.Image = image;
+            this.CompletionItemType = completionItemType;
+            this.FilePath = filePath;
+            this.ClassName = className;
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="CompletionData"/> class.
+        /// </summary>
+        /// <param name="text">The text of the completion item.</param>
+        /// <param name="description">A description of the completion item.</param>
+        /// <param name="image">An image associated with the completion item.</param>
+        /// <param name="completionItemType">The type of the completion item.</param>
+        /// <param name="filePath">The file path where the completion item is located.</param>
+        /// <param name="className">The name of the class associated with the completion item.</param>
+        /// <param name="methodName">The name of the method associated with the completion item.</param>
+        /// <param name="methodParameterTypes">A list of parameter types for the method.</param>
+        /// <param name="methodSignature">The method signature.</param>
+        /// <returns>
+        /// A new instance of the <see cref="CompletionData"/> class.
+        /// </returns>
+        public CompletionData(string text,
+                              string description,
+                              ImageSource image,
+                              CompletionItemType completionItemType,
+                              string filePath,
+                              string className,
+                              string methodName,
+                              List<string> methodParameterTypes,
+                              string methodSignature)
+        {
+            this.Text = text;
+            this.Description = description;
+            this.Image = image;
+            this.CompletionItemType = completionItemType;
+            this.FilePath = filePath;
+            this.ClassName = className;
+            this.MethodName = methodName;
+            this.MethodParameterTypes = methodParameterTypes;
+            this.MethodSignature = methodSignature;
+        }
+
+        /// <summary>
         /// Completes the text in the specified text area by replacing the content 
         /// within the given completion segment with the provided text.
         /// </summary>
@@ -88,7 +162,24 @@ namespace JeffPires.VisualChatGPTStudio.Utils.CodeCompletion
         /// <param name="insertionRequestEventArgs">Event arguments related to the insertion request.</param>
         public void Complete(TextArea textArea, ISegment completionSegment, EventArgs insertionRequestEventArgs)
         {
-            textArea.Document.Replace(completionSegment, this.Text + " ");
+            string text = this.Text + " ";
+
+            if (this.CompletionItemType == CompletionItemType.CSharpMethod)
+            {
+                text = this.Text + this.MethodSignature + " ";
+            }
+
+            textArea.Document.Replace(completionSegment, text);
         }
+    }
+
+    /// <summary>
+    /// Represents the different types of completion items available.
+    /// </summary>
+    public enum CompletionItemType
+    {
+        File = 0,
+        CSharpClass = 1,
+        CSharpMethod = 2
     }
 }